Added first version of edge graph generation function.
Brute force method first maps all the edges to a vertex in the new graph with a hash table, and then for every edge checks for incident edges. Those edges are looked up in the hash table and new edges are added between the two vertices that the original edges map to.